pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "astrospace"
version = "1.2.3"
description = "A Flask-based astrophotography web application"
authors = [
  { name="Sharon Shaji", email="sharonshaji92@outlook.com" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
"Flask",
"flask-login",
"flask_wtf",
"psycopg2-binary",
"numpy",
"pandas",
"requests",
"ephem",
"astroquery",
"astropy",
"photoutils",
"pillow",
"xisf",
"bleach",
"tinycss2"
]

[project.optional-dependencies]
dev = [
"pytest",
]

[tool.hatch.build]
include = [
  "AstroSpace/**/*.py",
  "AstroSpace/*.sql",
  "AstroSpace/static/**",
  "AstroSpace/templates/**",
  "AstroSpace/utils/**",
  "README.md",
  "LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["AstroSpace"]

[tool.pytest.ini_options]
testpaths = ["tests"]
